+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Add support for translating german to english
 
 [Unreleased]: https://github.com/bakku/dict/compare/e72d8a4...HEAD